# Optimize Scheduling when using StartAfter