--- a/tasks.go
+++ b/tasks.go
@@ -265,11 +265,7 @@ func (schd *Scheduler) AddWithID(id string, t *Task) error {
 	t.id = id
 	schd.tasks[t.id] = t
 
-	if t.StartAfter.IsZero() {
-		schd.scheduleTask(t)
-		return nil
-	}
-	go schd.scheduleTask(t)
+	schd.scheduleTask(t)
 	return nil
 }