This PR optimizes how tasks are scheduled when using the `StartAfter` option. Previously, the scheduler would create a goroutine when using `StartAfter` and wait for a response from `time.After()`. This change uses `time.AfterFunc()` to create a one-off task that then schedules another task.

Some data from benchmarks.

Before:
```
$ go test -run=Benchmark -bench ./...
goos: darwin
goarch: amd64
pkg: github.com/madflojo/tasks
cpu: Intel(R) Core(TM) i5-6360U CPU @ 2.00GHz
BenchmarkTasks/Adding_a_scheduler-4             219752        5418 ns/op       722 B/op       10 allocs/op
BenchmarkTasks/Looking_up_a_scheduled_task-4          35873827          32.75 ns/op        0 B/op        0 allocs/op
PASS
ok    github.com/madflojo/tasks 2.955s
```

After:
```
$ go test -run=Benchmark -bench ./...
goos: darwin
goarch: amd64
pkg: github.com/madflojo/tasks
cpu: Intel(R) Core(TM) i5-6360U CPU @ 2.00GHz
BenchmarkTasks/Adding_a_scheduler-4             635785        2245 ns/op       500 B/op        8 allocs/op
BenchmarkTasks/Looking_up_a_scheduled_task-4          35625927          32.69 ns/op        0 B/op        0 allocs/op
PASS
ok    github.com/madflojo/tasks 3.794s
```